Rename files and routes to follow consistent naming convention
